Move all vector `IntoParallelIterator` together
